refactor: reorganize component structure and apply best practices
