+++ src/app/page.tsx
@@ -1,5 +1,5 @@
 import { fetchBlogPosts } from "@/lib/api";
-import BlogList from "@/components/BlogList";
+import BlogList from "@/components/blog/BlogList";
 
 export default async function BlogPage() {
   const posts = await fetchBlogPosts();
